hypr: rehaul GUI apps placement, remove calls to $delay_hyprstart
delay_hyprstart was used because waybar was not able to connect peoperly
to display workspaces. but even after giving a 2 second delay (after
delay_hyprstart), it still doesn't work. So the idea is not correct.
Maybe a simpler one would be a waybar button/an alias to kill and
relaunch waybar